fix: move gptVersion from types to interfaces to solve gitignore
