+++ src/services/item/plugins/app/chatBot/interfaces/gptVersion.ts
@@ -1,0 +1,4 @@
+export enum GPTVersion {
+  GPT_3_5_TURBO = 'gpt-3.5-turbo',
+  GPT_4 = 'gpt-4',
+}
